+++ src/Monoid.ts
@@ -34,6 +34,7 @@
  */
 import { Bounded } from './Bounded'
 import { Endomorphism, getEndomorphismMonoid as getEM, getMonoid } from './function'
+import * as _ from './internal'
 import { ReadonlyRecord } from './ReadonlyRecord'
 import * as Se from './Semigroup'
 
@@ -139,9 +140,10 @@ export const reverse = <A>(M: Monoid<A>): Monoid<A> => ({
  */
 export const struct = <A>(monoids: { [K in keyof A]: Monoid<A[K]> }): Monoid<{ readonly [K in keyof A]: A[K] }> => {
   const empty: A = {} as any
-  // tslint:disable-next-line: forin
-  for (const key in monoids) {
-    empty[key] = monoids[key].empty
+  for (const k in monoids) {
+    if (_.hasOwnProperty.call(monoids, k)) {
+      empty[k] = monoids[k].empty
+    }
   }
   return {
     concat: Se.struct(monoids).concat,